Renamed "API", use getopts, add help.
Changed prefix from PBUI to PB.
Replaced array instantiation for literal in `_UIPasteboardInitialize`.
Fix an out-of-bounds write in `PBCreateBufferFromFd`.
Fix a segmentation fault in `PBCreateBufferFromFd` when writing to
`length`.
Add support for JPEG images.
Add support for overriding types on `PBPasteboardPerformCopy` and
`PBPasteboardPerformPaste`.
Add help text.
Add examples to README.